+++ src/components/cards/reward-card/components/title.js
@@ -17,11 +17,7 @@ var _propTypes = _interopRequireDefault(require("prop-types"));
 
 var _classnames = _interopRequireDefault(require("classnames"));
 
-var _text = require("../../../../components/typography/text");
-
-var _horizontalStroke = require("../../../../components/layout/horizontal-stroke");
-
-var _colorsConfig = _interopRequireDefault(require("../../../../constants/colors-config"));
+var _titleWithStroke = require("../../../../components/typography/title-with-stroke");
 
 var RewardCardTitle = function RewardCardTitle(_ref) {
   var children = _ref.children,
@@ -29,19 +25,15 @@ var RewardCardTitle = function RewardCardTitle(_ref) {
       disabled = _ref.disabled,
       className = _ref.className,
       props = (0, _objectWithoutProperties2.default)(_ref, ["children", "tagName", "disabled", "className"]);
-  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, /*#__PURE__*/_react.default.createElement(_text.Text, (0, _extends2.default)({}, props, {
-    fontStyle: "italic",
-    weight: "bold",
+  return /*#__PURE__*/_react.default.createElement(_titleWithStroke.TitleWithStroke, (0, _extends2.default)({}, props, {
+    italic: true,
+    margin: false,
     tag: tagName,
-    color: "font1",
+    modifier: "senary",
     className: (0, _classnames.default)('k-RewardCard__title', className, {
       'k-RewardCard__title--disabled': disabled
-    })
-  }), children), /*#__PURE__*/_react.default.createElement(_horizontalStroke.HorizontalStroke, {
-    size: "default",
-    className: "k-u-margin-bottom-triple",
-    color: disabled ? _colorsConfig.default.font2 : null
-  }));
+    }, 'k-u-color-font1')
+  }), children);
 };
 
 exports.RewardCardTitle = RewardCardTitle;